requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.3
mysql-connector-python==8.0.31
protobuf==3.20.1
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.6
schedule==1.1.0
tomli==2.0.1
typing_extensions==4.4.0
yarl==1.8.2
